Step 1 of migrating to class-based components, move App.js over.
Had been kinda avoiding this because (a) existing code was all as
functions, and (b) I'm not totally sure what the implications are for
browswer compatibility.
However, seems I can do it one piece at a time, so (a) isn't so bad.
As to (b), we'll cross that bridge if we ever come to it.
Until then, makes following guidelines from React's documentation a bit
easier.